# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  gamemanagement.CreateGameRoundRequest:
    properties:
      end_date:
        type: string
      event_id:
        type: string
      name:
        maxLength: 100
        type: string
      round_number:
        type: integer
      round_type:
        enum:
        - pool
        - bracket
        - semifinal
        - final
        type: string
      start_date:
        type: string
    required:
    - event_id
    - name
    - round_type
    type: object
...
  gamemanagement.GameRoundDTO:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      event_id:
        type: string
      games_count:
        type: integer
      id:
        type: string
      name:
        type: string
      round_number:
        type: integer
      round_type:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
    type: object
...
  gamemanagement.UpdateGameRoundRequest:
    properties:
      end_date:
        type: string
      name:
        maxLength: 100
        type: string
      round_number:
        type: integer
      round_type:
        enum:
        - pool
        - bracket
        - semifinal
        - final
        type: string
      start_date:
        type: string
    type: object
...
  handlers.ErrorResponse:
    description: Error response structure
    properties:
      error:
        example: error message
        type: string
    type: object
...
  handlers.PlayerStatResponse:
    properties:
      assists:
        type: integer
      gamesPlayed:
        type: integer
      goals:
        type: integer
      playerId:
        type: string
      playerName:
        type: string
      teamId:
        type: string
      teamName:
        type: string
    type: object
...
  handlers.SpiritBreakdownAverage:
    properties:
      attitude:
        type: number
      communication:
        type: number
      fairMindedness:
        type: number
      foulsBodyContact:
        type: number
      rulesKnowledge:
        type: number
    type: object
  handlers.SpiritLeaderboardResponse:
    properties:
      averageScore:
        type: number
      breakdown:
        $ref: '#/definitions/handlers.SpiritBreakdownAverage'
      gamesPlayed:
        type: integer
      teamId:
        type: string
      teamName:
        type: string
    type: object
...
  metadata.CountryDTO:
    properties:
      code:
        type: string
      continent_id:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for DigiGameStats reimplementation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DigiGameStats API
  version: "1.0"
paths:
...
  /events/{event_id}/rounds:
    get:
      description: List all rounds for an event
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/gamemanagement.GameRoundDTO'
            type: array
        "400":
          description: bad request
          schema:
            type: string
      summary: List Game Rounds
      tags:
      - game-rounds
    post:
      consumes:
      - application/json
      description: Create a new round for an event (pool, bracket, etc.)
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      - description: Game Round Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/gamemanagement.CreateGameRoundRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gamemanagement.GameRoundDTO'
        "400":
          description: bad request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Game Round
      tags:
      - game-rounds
...
  /geographic/countries:
    get:
      consumes:
      - application/json
      description: Get a list of all countries, optionally filtered by continent.
      parameters:
      - description: Filter by continent ID
        in: query
        name: continent_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/metadata.CountryDTO'
            type: array
      summary: List Countries
      tags:
      - geographic
...
  /public/leaderboards/players:
    get:
      description: Get top players by goals or assists
      parameters:
      - default: goals
        description: 'Category: goals or assists'
        in: query
        name: category
        type: string
      - description: Filter by event ID
        format: uuid
        in: query
        name: eventId
        type: string
      - description: Filter by division pool ID
        format: uuid
        in: query
        name: divisionPoolId
        type: string
      - default: 50
        description: Limit results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.PlayerStatResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get player leaderboard
      tags:
      - leaderboards
  /public/leaderboards/spirit:
    get:
      description: Get teams ranked by spirit scores
      parameters:
      - description: Filter by event ID
        format: uuid
        in: query
        name: eventId
        type: string
      - description: Filter by division pool ID
        format: uuid
        in: query
        name: divisionPoolId
        type: string
      - default: 50
        description: Limit results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.SpiritLeaderboardResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get spirit leaderboard
      tags:
      - leaderboards
...
  /rounds/{id}:
    get:
      description: Get game round details by ID
      parameters:
      - description: Round ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gamemanagement.GameRoundDTO'
        "404":
          description: not found
          schema:
            type: string
      summary: Get Game Round
      tags:
      - game-rounds
    put:
      consumes:
      - application/json
      description: Update game round details
      parameters:
      - description: Round ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/gamemanagement.UpdateGameRoundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gamemanagement.GameRoundDTO'
        "400":
          description: bad request
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Game Round
      tags:
      - game-rounds
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"